Revert "https://github.com/SummitShare/SummitShare-Client/pull/11"
