+++ packages/dapp/src/app/components/navigation/footer.tsx
@@ -65,9 +65,6 @@ function footer() {
                     Github
                   </a>
                 </li>
-                <li>
-                  <a href="https://t.me/+X3VvXwhsHnEwNTc0">Telegram</a>
-                </li>
               </ul>
             </li>
           </ul>
